Fix the event registry for Python3
inspect.ismethod behaves differently for Python 2 and 3 (and it seems that the
Py2 behaviour we're relying on here is actually a bug):

http://bugs.python.org/issue16851

I don't think we *need* to verify that the event is a method. As long as it has
the `_event_name` attribute (which we only set in the decorator), things should
just work.

Signed-off-by: Tomas Sedovic <tomas@sedovic.cz>